refactor(py_src): extract BeamWrapperGenerator into pipeline.Pipeline
The orchestrator class lived inside gen2.py alongside argparse glue,
mixing CLI plumbing with ~900 lines of class body. Move it to a
dedicated `pipeline.py` and rename to `Pipeline`; gen2.py becomes a
68-line CLI shim that just constructs `Pipeline` and calls `.gen()`.

Imports are scoped to what pipeline.Pipeline actually uses (no more
shotgun `from helper import *` at gen2's module top). The class body
itself is unchanged in this commit — splitting the 325-line gen()
into named phases is a follow-up.

Byte-equivalent vs lib-4.11.0/ and src-4.11.0/.
